requirements.txt
django_extensions
pep8
pyowm
